Add increments at user id column in musicians tab.

+++ database/migrations/2023_09_14_143505_create_musicians_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('musicians', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('user_id');
+            $table->increments('user_id');
             $table->date('birth_date');
             $table->string('address',70);
             $table->string('num_phone',50);
